simple refactoring for simple config :)
