+++ simple_config/__init__.py
@@ -1,0 +1,6 @@
+from .config import SimpleConfig
+
+
+__all__ = [
+    "SimpleConfig"
+]
